Final ver integration
