+++ src/main/java/tn/esprit/Controllers/Athletefront.java
@@ -15,11 +15,7 @@ public class Athletefront {
     @FXML
     private Label claimButton;
 
-    @FXML
-    private Button viewInjuryButton;
 
-    @FXML
-    private Button viewRecoveryPlanButton;
 
 
 
@@ -28,9 +24,7 @@ public class Athletefront {
         // Set action for claimButton
         claimButton.setOnMouseClicked(event -> switchScreenClaim2());
 
-        viewInjuryButton.setOnMouseClicked(event -> switchScreenListInjury());
 
-        viewRecoveryPlanButton.setOnMouseClicked(event -> switchScreenListRecoveryPlans());
 
 
     }
@@ -62,70 +56,6 @@ public class Athletefront {
         }
     }
 
-
-    private void switchScreenInjury() {
-        try {
-            FXMLLoader loader = new FXMLLoader(getClass().getResource("/AddAthleteInjury.fxml"));
-            Parent root = loader.load();
-
-            // Initialize the AddAthleteInjuryController for this screen
-            AddAthleteInjuryController controller = loader.getController();
-            controller.initialize();
-
-            Stage stage = new Stage();
-            stage.setScene(new Scene(root));
-            stage.setTitle("Add Injury");
-            stage.setUserData(this);
-            stage.show();
-
-            // Close the current window
-            Stage currentStage = (Stage) viewInjuryButton.getScene().getWindow();
-            currentStage.close();
-        } catch (IOException e) {
-            showAlert("Error", "Failed to open Add Injury screen: " + e.getMessage());
-        }
-    }
-
-    private void switchScreenListInjury() {
-        try {
-            FXMLLoader loader = new FXMLLoader(getClass().getResource("/ListInjury.fxml"));
-            Parent root = loader.load();
-
-            Stage stage = new Stage();
-            stage.setScene(new Scene(root));
-            stage.setTitle("Injury List");
-            stage.setUserData(this);
-            stage.show();
-
-            // Close the current window
-            Stage currentStage = (Stage) viewInjuryButton.getScene().getWindow();
-            currentStage.close();
-        } catch (IOException e) {
-            showAlert("Error", "Failed to open Injury List screen: " + e.getMessage());
-        }
-    }
-
-
-    private void switchScreenListRecoveryPlans() {
-        try {
-            FXMLLoader loader = new FXMLLoader(getClass().getResource("/ListRecoveryPlan.fxml"));
-            Parent root = loader.load();
-
-            Stage stage = new Stage();
-            stage.setScene(new Scene(root));
-            stage.setTitle("Recovery Plans List");
-            stage.setUserData(this);
-            stage.show();
-
-            // Close the current window
-            Stage currentStage = (Stage) viewRecoveryPlanButton.getScene().getWindow();
-            currentStage.close();
-        } catch (IOException e) {
-            e.printStackTrace(); // Print the error in the console
-            showAlert("Error", "Failed to open Recovery Plans List screen: " + e.getMessage());
-        }
-
-    }
 
 
 }
