+++ src/main/resources/Medicalfront.fxml
@@ -1,14 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.effect.DropShadow?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
-<?import javafx.scene.paint.Color?>
 <?import javafx.scene.text.Font?>
 
 <AnchorPane prefHeight="412.0" prefWidth="600.0" style="-fx-background-color: #709775;" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="tn.esprit.controllers.Medicalfront">
@@ -25,7 +24,8 @@
                                 <Insets left="40.0" top="5.0" />
                             </VBox.margin>
                         </ImageView>
-                    </children></VBox>
+                    </children>
+                </VBox>
                 <HBox alignment="CENTER" prefHeight="38.0" prefWidth="157.0">
                     <children>
                         <Label prefHeight="50.0" prefWidth="75.0" text="Athletes" textFill="WHITE">
@@ -39,49 +39,54 @@
                     </VBox.margin>
                 </HBox>
                 <HBox alignment="CENTER" prefHeight="38.0" prefWidth="157.0">
-                    <children>
-                        <Label prefHeight="50.0" prefWidth="59.0" text="Injury" textFill="WHITE">
+                    <VBox.margin>
+                        <Insets top="5.0" />
+                    </VBox.margin>
+               <children>
+                        <Button fx:id="viewRecoveryPlanButton" prefHeight="38.0" prefWidth="191.0" style="-fx-background-color: #709775; -fx-text-fill: white; -fx-font-size: 18px; -fx-font-weight: bold; -fx-border-radius: 8px; -fx-background-radius: 8px; -fx-padding: 5px 10px;" text="Recovery Plans">
+                            <font>
+                                <Font name="System Bold" size="16.0" />
+                            </font>
+                        </Button>
+               </children>
+                </HBox>
+                <HBox alignment="CENTER" prefHeight="38.0" prefWidth="157.0">
+                    <VBox.margin>
+                        <Insets top="5.0" />
+                    </VBox.margin>
+               <children>
+                        <Button fx:id="viewInjuryButton" prefHeight="38.0" prefWidth="176.0" style="-fx-background-color: #709775; -fx-text-fill: white; -fx-font-size: 18px; -fx-font-weight: bold; -fx-border-radius: 8px; -fx-background-radius: 8px; -fx-padding: 5px 10px;" text="Injuries">
+                            <font>
+                                <Font name="System Bold" size="16.0" />
+                            </font>
+                        </Button>
+               </children>
+                </HBox>
+                <HBox alignment="CENTER" prefHeight="38.0" prefWidth="157.0">
+                    <VBox.margin>
+                        <Insets top="5.0" />
+                    </VBox.margin>
+               <children>
+                        <Label prefHeight="38.0" prefWidth="62.0" text="Teams" textFill="WHITE">
                             <font>
                                 <Font name="Calibri Bold" size="20.0" />
                             </font>
                         </Label>
-                    </children>
+               </children>
+                </HBox>
+                <HBox alignment="CENTER" prefHeight="38.0" prefWidth="157.0">
                     <VBox.margin>
                         <Insets top="5.0" />
                     </VBox.margin>
-                </HBox>
-                <HBox alignment="CENTER" prefHeight="38.0" prefWidth="157.0">
-                    <children>
-                        <Label prefHeight="38.0" prefWidth="122.0" text="Recovery Plan" textFill="WHITE">
-                            <font>
-                                <Font name="Calibri Bold" size="20.0" />
-                            </font>
-                        </Label>
-                    </children>
-                    <VBox.margin>
-                        <Insets top="5.0" />
-                    </VBox.margin>
-                </HBox>
-                <HBox alignment="CENTER" prefHeight="38.0" prefWidth="157.0">
-                    <children>
-                        <Label prefHeight="38.0" prefWidth="123.0" text="Nutrition Plan" textFill="WHITE">
-                            <font>
-                                <Font name="Calibri Bold" size="20.0" />
-                            </font>
-                        </Label>
-                    </children>
-                    <VBox.margin>
-                        <Insets top="5.0" />
-                    </VBox.margin>
-                </HBox>
-                <HBox alignment="CENTER" prefHeight="38.0" prefWidth="157.0">
-                    <children>
+               <children>
                         <Label fx:id="claimButton" prefHeight="38.0" prefWidth="53.0" text="Claim" textFill="WHITE">
                             <font>
                                 <Font name="Calibri Bold" size="20.0" />
                             </font>
                         </Label>
-                    </children>
+               </children>
+                </HBox>
+                <HBox alignment="CENTER" prefHeight="38.0" prefWidth="157.0">
                     <VBox.margin>
                         <Insets top="5.0" />
                     </VBox.margin>
@@ -139,27 +144,6 @@
                                             </children>
                                         </VBox>
                                     </children>
-                                </HBox>
-                                <ImageView fitHeight="97.0" fitWidth="117.0" pickOnBounds="true" preserveRatio="true">
-                                    <image>
-                                        <Image url="@vector.png" />
-                                    </image>
-                                </ImageView>
-                            </children>
-                        </HBox>
-                        <HBox prefHeight="260.0" prefWidth="439.0">
-                            <children>
-                                <HBox prefHeight="260.0" prefWidth="431.0" style="-fx-background-color: #f9f9f9; -fx-background-radius: 30;">
-                                    <HBox.margin>
-                                        <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-                                    </HBox.margin>
-                                    <effect>
-                                        <DropShadow height="7.0" radius="3.0" width="7.0">
-                                            <color>
-                                                <Color red="0.14473684132099152" green="0.12329434603452682" blue="0.12329434603452682" />
-                                            </color>
-                                        </DropShadow>
-                                    </effect>
                                 </HBox>
                             </children>
                         </HBox>
